feat: photoSpot api - 스팟 조회, 상세 조회

+++ src/main/java/app/gangdan/please/domain/photoSpot/PhotoSpotCustomRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 
 public interface PhotoSpotCustomRepository {
 
-    List<PhotoSpot> getSpotList(double latitude, double longitude, Long radius);
+    List<PhotoSpot> getSpotList(double latitude, double longitude, double radius);
 
 }
